Refactor of menu code
